+++ etc/script/oph_start.sh
@@ -31,6 +31,7 @@ serverid=${6}
 workflowid=${7}
 project=${8}
 taskname=${9}
+ndbms=${10}
 
 # Const
 fixString=