+++ src/oph_parser.h
@@ -27,6 +27,7 @@
 #define OPH_ARG_WORKFLOWID "workflowid"
 #define OPH_ARG_NCORES "ncores"
 #define OPH_ARG_NHOSTS "nhost"
+#define OPH_ARG_NDBMS "ndbms"
 #define OPH_ARG_MODE "exec_mode"
 #define OPH_ARG_OPERATOR "operator"
 #define OPH_ARG_USER "user"